refactor: `start()` should throw if filename doesn't exist
BREAKING CHANGE: Error handling for `start` is changing.

Since `start` is an `async` function, one would expect to
be able to try/catch it. The original version ran any
failures through the `quit` method, but it's designed to
fail gracefully and work asynchronously by using `'error'`
events. This causes weird flow issues if  `start` is throwing·
because the filename does not exist. This change allows
`start` to simply throw.

Semver: major
Ref: LOG-8030