fix: Specify 'r' flag for fs.promises.open
On older version of node (e.g. 10), the read flags cannot
be `undefined` or it will throw an error.  This change
specifies the 'r' flag, which defaults correctly in later
node versions.  Also, since this package uses for-await-of
loops when reading streams, version >= 10.3.0 is required.
This change specifies an `engines` declaration in the
package.json.

Semver: patch
Fixes: https://github.com/logdna/tail-file-node/issues/10