test: Add tap-xunit and flatten results for junit output
This change tests `tap-xunit` as a way to generate a junit
report. `tap-xunit` does not parse failures on subtests
properly, so flatten them in order to see the full failure
details which are important. The `tap-mocha-reporter`, which
we normally use, does not properly include the failure details,
so `tap-xunit` is a viable replacement.

Semver: patch