fix: Error on rotated file if new file takes time to appear
When a file's inode changes, or the `stat` call throws
causes tail-file to read the remainder of the contents
via a FileHandle. If the file disappears, but takes a bit
to re-appear, `_readRemainderFromFileHandle()` is
erroneously called twice; once for the `ENOENT` from the
disappearing file, and once when the new file causes the
inode to change. This commit fixes a bug where the FH is
nullified after the first call, and thus throws on the
subsequent call.

Semver: patch
Fixes: https://github.com/logdna/tail-file-node/issues/18