fix: Necessary changes for asyncIterators in recent version of node
Since node:16, `asyncIterator` has added an option called
`destroyOnReturn` which will abort the stream if an iterator loop does
not cleanly fall through. In the case of `TailFile`, backpressure is
respected, and that iteration will return early causing an `ABORT_ERR`
to be thrown.  The option for `destroyOnReturn` is true by default, so
this change turns off that experimental feature.

Semver: minor