chore(deps)!: Change `commitlint` tooling and add `semantic-release`
This commit changes to our internal tooling for `commitlint`, which
is better suited to enforce commit style for public PRs. This
also bumps eslint and its associated dependencies, and switches
to `semantic-release` tooling.

BREAKING CHANGE: `eslint@8.x` and other dependencies no longer support
node:10, so this change officially removes support for node:10 as
we can no longer do regression testing for it.

Fixes: #28